backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-dotenv==1.2.1
pymysql==1.1.1
bcrypt==4.3.0
reportlab==4.4.1
httpx==0.28.1
